pyproject.toml
[project]
name = "poker-card-detection"
version = "0.1.0"
description = "Real-time playing card detection and poker hand classification"
requires-python = ">=3.11"
dependencies = [
    "ultralytics>=8.3.0",
    "pytorch-lightning>=2.3.0",
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "mlflow>=2.14.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "streamlit>=1.36.0",
    "dvc>=3.51.0",
    "fire>=0.6.0",
    "roboflow>=1.1.40",
    "gitpython>=3.1.43",
    "python-multipart>=0.0.9",
    "onnx>=1.16.0",
    "onnxruntime>=1.18.0",
    "onnxslim>=0.1.71",
    "tritonclient[http]>=2.40.0",
    "opencv-python-headless>=4.10.0",
    "numpy>=1.26.0",
    "Pillow>=10.4.0",
]

[project.scripts]
poker-cards = "poker_card_detection.commands:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["poker_card_detection"]

[dependency-groups]
dev = [
    "pre-commit>=3.7.0",
    "pytest>=8.2.0",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cu128", marker = "sys_platform == 'win32'" },
]
torchvision = [
    { index = "pytorch-cu128", marker = "sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.isort]
profile = "black"
line_length = 88

[tool.black]
line-length = 88
target-version = ["py311"]
